+++ crates/sage-cloudpath/src/util.rs
@@ -111,6 +111,16 @@ pub fn read_mgf<S: AsRef<str>>(path: S, file_id: usize) -> Result<Vec<RawSpectru
     })
 }
 
+pub fn read_text<S: AsRef<str>>(path: S) -> Result<String, Error> {
+    read_and_execute(path, |mut bf| async move {
+        let mut contents = String::new();
+        bf.read_to_string(&mut contents)
+            .await
+            .map_err(crate::Error::IO)?;
+        Ok(contents)
+    })
+}
+
 pub fn read_fasta<S>(
     path: S,
     decoy_tag: S,